requirements.txt
python-dotenv
whisperx
openai
genanki
anki